requirements.txt
torch
torchvision
einops
numpy
Pillow
psutil
